# Improve performance of Time to/from JSON