--- a/model/time_test.go
+++ b/model/time_test.go
@@ -131,14 +131,14 @@ func TestParseDuration(t *testing.T) {
 	}
 }
 
-func BenchmarkMarshalJSON(b *testing.B) {
+func BenchmarkTimeMarshalJSON(b *testing.B) {
 	t := TimeFromUnixNano(1514834431600000000)
 	for i := 0; i < b.N; i++ {
 		t.MarshalJSON()
 	}
 }
 
-func BenchmarkUnmarshalJSON(b *testing.B) {
+func BenchmarkTimeUnmarshalJSON(b *testing.B) {
 	t := TimeFromUnixNano(1514834431600000000)
 	buf, _ := t.MarshalJSON()
 	for i := 0; i < b.N; i++ {