A few changes to speed up marshalling/unmarshalling of `model.Time` - see individual commit messages.

Benchmarks: before:
```
BenchmarkMarshalJSON-2     	 5000000	       303 ns/op	      64 B/op	       3 allocs/op
BenchmarkUnmarshalJSON-2   	 5000000	       314 ns/op	      53 B/op	       5 allocs/op
```

after:
```
BenchmarkMarshalJSON-2     	10000000	       219 ns/op	      16 B/op	       1 allocs/op
BenchmarkUnmarshalJSON-2   	10000000	       124 ns/op	      16 B/op	       2 allocs/op
```

@fabxc