@bboreham 2018-03-09
Note prometheus/prometheus#3536 overrides `MarshalJSON` in `Point` so this PR has less impact.  I still have cases where I'm marshalling `Time` directly.

#3536 doesn't touch the `UnmarshalJSON` side.