@bboreham 2018-03-21
Aha, hadn't picked up on that shift.

Used here: https://github.com/weaveworks/cortex/blob/cc8cdf499162077c4a21ae05d162ca8116228718/pkg/chunk/chunk.go#L44-L45
no biggy to change it to underlying types.