@brian-brazil 2018-03-21
Changing to underlying types is probably best, it'll be less of a problem when this code eventually goes away.