@bboreham 2018-04-06
In the end I moved my code to use `jsoniter` and added type overrides to achieve the same effect.

You might consider a note in the README or CONTRIBUTING here that `model` is deprecated.